Add files via upload
Q1.	Take a square matrix of 4x4 and filled it randomly with different numbers by any readymade function. Compute the following requirements:
(a)	Mean of Diagonal elements of square matrix
(b)	Mean of upper part of diagonal of square matrix
(c)	Mean of lower part of diagonal of square matrix
(d)	The sum of the cells of square matrix such that the sum of the index should be divisible by 2